api/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-dotenv>=1.0
pydantic>=2.0
pydantic-settings>=2.0
python-multipart>=0.0.9
aiofiles>=23.0
